django_app/requirements.txt
defusedxml==0.5.0
Django==1.10.6
django-cors-headers==2.0.2
django-filter==1.0.2
django-rest-auth==0.9.1
djangorestframework==3.6.2
oauthlib==2.0.2
olefile==0.44
Pillow==4.0.0
Pygments==2.2.0
PyJWT==1.4.2
python-social-auth==0.3.6
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-core==1.2.0
